Make suggested changes
